requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
kiwisolver==1.2.0
matplotlib==3.3.1
networkx==2.5
numpy==1.19.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.8.2
six==1.15.0
texttable==1.6.3
